Uint texture indices

+++ assets/shaders/component.vert
@@ -3,8 +3,7 @@
 layout(location = 0) in vec2 a_position;
 layout(location = 1) in mat3 a_world;
 layout(location = 4) in mat3 a_texture;
-/** @todo Replace by uint? */
-layout(location = 7) in float a_textureIndex;
+layout(location = 7) in uint a_textureIndex;
 
 uniform mat3 u_projection;
 
@@ -16,6 +15,6 @@ void main() {
 
 	gl_Position = vec4(u_projection * a_world * position, 1);
 
-	v_textureIndex = uint(a_textureIndex);
+	v_textureIndex = a_textureIndex;
 	v_uv = (a_texture * position).xy;
 }
